[#223] Moved SubmissionAttachment tests to own file
